created llm.py and test.py files
llm.py loads a dataset and creates a model by fine tuning the pre-trained model (distilbert-base-uncased) with the loaded dataset (currently uses squad dataset from hugging face). Should not have to run this file again since the model has already been pushed to hugging face hub, unless more fine tuning needs to be done with another dataset.
test.py is a test file that takes in a question, reads a pdf from a directory path (will need to change later to pull document from database), extracts the text, and uses the text as the context for question and answering. Currently, the answer retrieved from the context is incorrect.